把col设为uint64 编译通过

+++ src/libfm/src/Data.h
@@ -88,7 +88,7 @@ class Data {
 		LargeSparseMatrix<DATA_FLOAT>* data;
 		DVector<DATA_FLOAT> target;
 
-		int num_feature;
+		uint64 num_feature;
 		uint num_cases;
  
 		DATA_FLOAT min_target;
